+++ resilience4j-annotations/README.adoc
@@ -1,0 +1,13 @@
+= resilience4j-annotations
+
+Annotations to be implemented for applying resilience4j elements in DI projects (Spring, CDI, etc)
+
+== License
+
+Copyright 2018 Robert Winkler and Raffaele Litto
+
+Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
