+++ resilience4j-spring-boot/src/main/java/io/github/resilience4j/circuitbreaker/autoconfigure/CircuitBreakerAutoConfiguration.java
@@ -15,21 +15,16 @@
  */
 package io.github.resilience4j.circuitbreaker.autoconfigure;
 
-import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+import io.github.resilience4j.circuitbreaker.CircuitBreaker;
+import io.github.resilience4j.circuitbreaker.CircuitBreakerRegistry;
+import io.github.resilience4j.circuitbreaker.configure.CircuitBreakerConfiguration;
+import io.github.resilience4j.circuitbreaker.configure.CircuitBreakerProperties;
+import io.github.resilience4j.circuitbreaker.monitoring.endpoint.CircuitBreakerEndpoint;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import io.github.resilience4j.circuitbreaker.CircuitBreaker;
-import io.github.resilience4j.circuitbreaker.CircuitBreakerConfig;
-import io.github.resilience4j.circuitbreaker.CircuitBreakerRegistry;
-import io.github.resilience4j.circuitbreaker.event.CircuitBreakerEvent;
-import io.github.resilience4j.circuitbreaker.internal.InMemoryCircuitBreakerRegistry;
-import io.github.resilience4j.circuitbreaker.monitoring.endpoint.CircuitBreakerEndpoint;
-import io.github.resilience4j.circuitbreaker.monitoring.health.CircuitBreakerHealthIndicator;
-import io.github.resilience4j.consumer.DefaultEventConsumerRegistry;
-import io.github.resilience4j.consumer.EventConsumerRegistry;
+import org.springframework.context.annotation.Import;
 
 /**
  * {@link org.springframework.boot.autoconfigure.EnableAutoConfiguration
@@ -37,50 +32,12 @@ import io.github.resilience4j.consumer.EventConsumerRegistry;
  */
 @Configuration
 @ConditionalOnClass(CircuitBreaker.class)
+@Import(CircuitBreakerConfiguration.class)
 @EnableConfigurationProperties(CircuitBreakerProperties.class)
 public class CircuitBreakerAutoConfiguration {
-
-    @Bean
-    public CircuitBreakerRegistry circuitBreakerRegistry(CircuitBreakerProperties circuitBreakerProperties,
-                                                         EventConsumerRegistry<CircuitBreakerEvent> eventConsumerRegistry,
-                                                         ConfigurableBeanFactory beanFactory) {
-        CircuitBreakerRegistry circuitBreakerRegistry = new InMemoryCircuitBreakerRegistry();
-        circuitBreakerProperties.getBackends().forEach(
-            (name, properties) -> {
-                CircuitBreakerConfig circuitBreakerConfig = circuitBreakerProperties.createCircuitBreakerConfig(name);
-                CircuitBreaker circuitBreaker = circuitBreakerRegistry.circuitBreaker(name, circuitBreakerConfig);
-                circuitBreaker.getEventPublisher().onEvent(eventConsumerRegistry.createEventConsumer(name, properties.getEventConsumerBufferSize()));
-
-                if (properties.getRegisterHealthIndicator()) {
-                    CircuitBreakerHealthIndicator healthIndicator = new CircuitBreakerHealthIndicator(circuitBreaker);
-                    beanFactory.registerSingleton(
-                        name + "CircuitBreakerHealthIndicator",
-                        healthIndicator
-                    );
-                }
-            }
-        );
-        return circuitBreakerRegistry;
-    }
-
-    @Bean
-    public CircuitBreakerAspect circuitBreakerAspect(CircuitBreakerProperties circuitBreakerProperties,
-                                                     CircuitBreakerRegistry circuitBreakerRegistry) {
-        return new CircuitBreakerAspect(circuitBreakerProperties, circuitBreakerRegistry);
-    }
-
     @Bean
     public CircuitBreakerEndpoint circuitBreakerEndpoint(CircuitBreakerRegistry circuitBreakerRegistry) {
         return new CircuitBreakerEndpoint(circuitBreakerRegistry);
     }
 
-    /**
-     * The EventConsumerRegistry is used to manage EventConsumer instances.
-     * The EventConsumerRegistry is used by the CircuitBreakerHealthIndicator to show the latest CircuitBreakerEvents events
-     * for each CircuitBreaker instance.
-     */
-    @Bean
-    public EventConsumerRegistry<CircuitBreakerEvent> eventConsumerRegistry() {
-        return new DefaultEventConsumerRegistry<>();
-    }
-}
+ }
