+++ resilience4j-spring-boot/src/main/java/io/github/resilience4j/ratelimiter/autoconfigure/RateLimiterMetricsAutoConfiguration.java
@@ -15,19 +15,15 @@
  */
 package io.github.resilience4j.ratelimiter.autoconfigure;
 
-import com.codahale.metrics.MetricRegistry;
-
+import io.github.resilience4j.ratelimiter.configure.RateLimiterMetricsConfiguration;
 import org.springframework.boot.actuate.autoconfigure.MetricRepositoryAutoConfiguration;
 import org.springframework.boot.actuate.autoconfigure.MetricsDropwizardAutoConfiguration;
 import org.springframework.boot.actuate.metrics.repository.MetricRepository;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import io.github.resilience4j.metrics.RateLimiterMetrics;
-import io.github.resilience4j.ratelimiter.RateLimiterRegistry;
+import org.springframework.context.annotation.Import;
 
 /**
  * {@link org.springframework.boot.autoconfigure.EnableAutoConfiguration
@@ -35,13 +31,10 @@ import io.github.resilience4j.ratelimiter.RateLimiterRegistry;
  */
 @Configuration
 @ConditionalOnClass(MetricRepository.class)
+@Import(RateLimiterMetricsConfiguration.class)
+
 @AutoConfigureAfter(value = {RateLimiterAutoConfiguration.class, MetricsDropwizardAutoConfiguration.class})
 @AutoConfigureBefore(MetricRepositoryAutoConfiguration.class)
 public class RateLimiterMetricsAutoConfiguration {
-    @Bean
-    public RateLimiterMetrics registerRateLimiterMetrics(RateLimiterRegistry rateLimiterRegistry, MetricRegistry metricRegistry) {
-        RateLimiterMetrics rateLimiterMetrics = RateLimiterMetrics.ofRateLimiterRegistry(rateLimiterRegistry);
-        metricRegistry.registerAll(rateLimiterMetrics);
-        return rateLimiterMetrics;
-    }
+
 }
