+++ resilience4j-spring/build.gradle
@@ -1,17 +1,12 @@
 dependencies {
-    compile ( libraries.spring_boot_aop )
-    compile ( libraries.spring_boot_actuator )
-    compile ( libraries.spring_boot_web )
-    compile ( libraries.spring_reactor )
+    compile ( libraries.aspectj )
+    compile ( libraries.spring_4_core, libraries.spring_4_context )
     compile project(':resilience4j-annotations')
-    compile project(':resilience4j-spring')
+    compile project(':resilience4j-consumer')
     compile project(':resilience4j-circuitbreaker')
     compile project(':resilience4j-ratelimiter')
-    compile project(':resilience4j-consumer')
     compileOnly project(':resilience4j-prometheus')
     compileOnly project(':resilience4j-metrics')
-    testCompile ( libraries.spring_boot_test )
     testCompile project(':resilience4j-prometheus')
     testCompile project(':resilience4j-metrics')
-    testCompile ( libraries.prometheus_spring_boot )
 }
