+++ resilience4j-spring/src/main/java/io/github/resilience4j/circuitbreaker/configure/CircuitBreakerMetricsConfiguration.java
@@ -13,26 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.resilience4j.circuitbreaker.autoconfigure;
+package io.github.resilience4j.circuitbreaker.configure;
 
-import io.github.resilience4j.circuitbreaker.configure.CircuitBreakerMetricsConfiguration;
+import com.codahale.metrics.MetricRegistry;
+import io.github.resilience4j.circuitbreaker.CircuitBreakerRegistry;
+import io.github.resilience4j.metrics.CircuitBreakerMetrics;
 import org.springframework.boot.actuate.autoconfigure.MetricRepositoryAutoConfiguration;
 import org.springframework.boot.actuate.autoconfigure.MetricsDropwizardAutoConfiguration;
 import org.springframework.boot.actuate.metrics.repository.MetricRepository;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 
-/**
- * {@link org.springframework.boot.autoconfigure.EnableAutoConfiguration
- * Auto-configuration} for resilience4j-metrics.
- */
 @Configuration
-@ConditionalOnClass(MetricRepository.class)
-@Import(CircuitBreakerMetricsConfiguration.class)
-@AutoConfigureAfter(value = {CircuitBreakerAutoConfiguration.class, MetricsDropwizardAutoConfiguration.class})
-@AutoConfigureBefore(MetricRepositoryAutoConfiguration.class)
-public class CircuitBreakerMetricsAutoConfiguration {
+public class CircuitBreakerMetricsConfiguration {
+
+    @Bean
+    public CircuitBreakerMetrics registerCircuitBreakerMetrics(CircuitBreakerRegistry circuitBreakerRegistry, MetricRegistry metricRegistry){
+        CircuitBreakerMetrics circuitBreakerMetrics = CircuitBreakerMetrics.ofCircuitBreakerRegistry(circuitBreakerRegistry);
+        metricRegistry.registerAll(circuitBreakerMetrics);
+        return circuitBreakerMetrics;
+    }
 }
