+++ resilience4j-spring/src/main/java/io/github/resilience4j/ratelimiter/configure/RateLimiterPrometheusConfiguration.java
@@ -13,29 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.resilience4j.ratelimiter.autoconfigure;
+package io.github.resilience4j.ratelimiter.configure;
 
 import io.github.resilience4j.prometheus.RateLimiterExports;
 import io.github.resilience4j.ratelimiter.RateLimiterRegistry;
-import io.github.resilience4j.ratelimiter.configure.RateLimiterConfiguration;
-import io.github.resilience4j.ratelimiter.configure.RateLimiterPrometheusConfiguration;
-import org.springframework.boot.autoconfigure.AutoConfigureAfter;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 
 /**
- * {@link org.springframework.boot.autoconfigure.EnableAutoConfiguration
- * Auto-configuration} for resilience4j-metrics.
+ * {@link org.springframework.context.annotation.Configuration
+ * Configuration} for resilience4j-metrics.
  */
 @Configuration
-@Import(RateLimiterPrometheusConfiguration.class)
-@AutoConfigureAfter(value = RateLimiterAutoConfiguration.class)
-@ConditionalOnClass(RateLimiterExports.class)
-public class RateLimiterPrometheusAutoConfiguration {
+public class RateLimiterPrometheusConfiguration {
     @Bean
-    public RateLimiterExports rateLimiterPrometheusCollector(RateLimiterRegistry rateLimiterRegistry){
+    public RateLimiterExports rateLimiterPrometheusCollector(RateLimiterRegistry rateLimiterRegistry) {
         RateLimiterExports collector = RateLimiterExports.ofRateLimiterRegistry(rateLimiterRegistry);
         collector.register();
         return collector;
