+++ settings.gradle
@@ -12,6 +12,7 @@ include 'resilience4j-circularbuffer'
 include 'resilience4j-metrics'
 include 'resilience4j-consumer'
 include 'resilience4j-vertx'
+include 'resilience4j-annotations'
 include 'resilience4j-spring-boot'
 include 'resilience4j-spring-boot2'
 include 'resilience4j-retrofit'
@@ -21,4 +22,5 @@ include 'resilience4j-timelimiter'
 include 'resilience4j-rxjava2'
 include 'resilience4j-reactor'
 include 'resilience4j-micrometer'
+include 'resilience4j-spring'
